Fix temporal_degrain to pass sad_mode and use REPAIR for post processing. (#99)
Finally got some free time, and after a lot of comparison with my original TemporalDegrain2
implementation, I nailed down the primary sources of differences between
the two.

Essentially, sad_mode wasn't being passed in to the original MVTools
preset like it used to, and the post processing denoiser was flipped to
use (strong) DFTTest instead of the previous (much much lighter)
Repair/RemoveGrain.

There's still some *very* small differences, but it seems to be due to
some defaults buried deep down in the MVTools implementation, as well as
a minor difference in the FFT3DFilter parameters.

But with this, the two outputs are pretty pretty darn close, and are
good enough in my book.