Fix TRT selection in dpir
The `cuda is None` check is useless because `cuda` is always defined by
line 73 after the 0b3a105 refactor (this is caught by mypy). So what
actually ends up happening is that the TRT backend is always chosen if
it's supported by the current system, even if the user explicitly wants
another backend.

Before:

    dpir(clip)
    # backend = Backend.TRT

    dpir(clip, cuda=False)
    # backend = Backend.TRT

    dpir(clip, cuda=True)
    # backend = Backend.TRT

    dpir(clip, cuda="trt")
    # backend = Backend.TRT

After:

    dpir(clip)
    # backend = Backend.TRT

    dpir(clip, cuda=False)
    # backend = Backend.NCNN_VK

    dpir(clip, cuda=True)
    # backend = Backend.ORT_CUDA

    dpir(clip, cuda="trt")
    # backend = Backend.TRT