BM3D: Select class based on GPU available
Check if a GPU is available, and don't select any CUDA classes if it isn't. This should prevent users who don't have GPUs from running into errors because it auto-selected a BM3D version that requires a GPU.

Perhaps an additional check should be done to ensure the GPU supports cuda, but that can be done in a future commit.